Add a verification step to the orienation lock
The current generation clicks everything to test it out. When enabling an option they did not intend, developers are accused of releasing broken apps. Rather than review the instructions, this results in a hostile message. We are too late to raise these children right, so the next best thing is verify the erratic choices.